requirements.txt
asgiref==3.4.1
Django==3.2.8
django-nose==1.4.7
djangorestframework==3.12.4
GeoAlchemy2==0.9.4
greenlet==1.1.2
importlib-metadata==4.8.1
nose==1.3.7
packaging==21.0
psycopg2==2.9.1
pyparsing==2.4.7
pytz==2021.3
Shapely==1.7.1
SQLAlchemy==1.4.25
sqlparse==0.4.2
typing-extensions==3.10.0.2
zipp==3.6.0
